backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
pyjwt[crypto]==2.8.0
pydantic-settings==2.5.2
python-telegram-bot==21.0.1
supabase==2.4.4
tenacity==8.5.0
pydub==0.25.1
aiohttp==3.9.1
grpcio==1.67.0
grpcio-tools==1.67.0
protobuf==5.28.3
cryptography==42.0.5
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.17
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0